refactor: @Data 제거 및 가독성 위한 이름 변경
