--- src/main/java/com/meta/cloud/dto/auth/JoinResponse.java
@@ -1,15 +1,0 @@
-package com.meta.cloud.dto.auth;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-public class JoinResponse {
-
-    private UUID userId;
-
-    public JoinResponse(UUID userId) {
-        this.userId = userId;
-    }
-}
